+++ api/src/payments/globee/mod.rs
@@ -44,7 +44,7 @@ impl PaymentProcessor for GlobeePaymentProcessor {
     fn partial_refund(
         &self,
         _auth_token: &str,
-        _amount: u32,
+        _amount: i64,
     ) -> Result<ChargeAuthResult, PaymentProcessorError> {
         Err(PaymentProcessorError {
             description: "Refunds are not supported by this gateway".to_string(),
